organising solvers into their own files
